Move Deserializer to serialization package
