+++ domain/src/main/java/com/aconno/acnsensa/domain/serialization/Deserializer.kt
@@ -1,4 +1,6 @@
-package com.aconno.acnsensa.domain.format
+package com.aconno.acnsensa.domain.serialization
+
+import com.aconno.acnsensa.domain.format.ByteFormat
 
 interface Deserializer {
